+++ src/daf/client.py
@@ -108,6 +108,7 @@ class ACCOUNT:
         "intents",
         "_running",
         "tasks",
+        "_ws_task",
         "_servers",
         "_autoguilds",
         "_selenium",
@@ -164,6 +165,7 @@ class ACCOUNT:
 
         self._client = None
         self._deleted = False
+        self._ws_task = None
         attributes.write_non_exist(self, "_update_sem", asyncio.Semaphore(1))
 
     def __str__(self) -> str:
@@ -284,12 +286,14 @@ class ACCOUNT:
 
         # Login
         trace("Logging in...")
-        _client_task = asyncio.create_task(self._client.start(self._token, bot=not self.is_user))
         try:
+            await self._client.login(self._token, not self.is_user)
+            ws_task = asyncio.create_task(self._client.connect())
+            self._ws_task = ws_task
             await self._client.wait_for("ready", timeout=LOGIN_TIMEOUT_S)
             trace(f"Logged in as {self._client.user.display_name}")
         except asyncio.TimeoutError as exc:
-            exc = _client_task.exception() if _client_task.done() else exc
+            exc = ws_task.exception() if ws_task.done() else exc
             raise RuntimeError(f"Error logging in to Discord. (Token {self._token[:TOKEN_MAX_PRINT_LEN]}...)") from exc
 
         for server in self._uiservers:
@@ -447,6 +451,7 @@ class ACCOUNT:
             await guild_._close()
 
         await self._client.close()
+        await asyncio.gather(self._ws_task, return_exceptions=True)
         self.client.clear()
 
         if _delete: